requirement.txt
absl-py==1.0.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.10.8
cleverhans==3.0.1
cloudpickle==2.0.0
cycler==0.11.0
decorator==5.1.1
dm-tree==0.1.6
fonttools==4.28.5
gast==0.2.2
google-pasta==0.2.0
grpcio==1.43.0
h5py==2.10.0
importlib-metadata==4.10.0
joblib==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.6
matplotlib==3.5.1
mnist==0.2.2
nose==1.3.7
numpy==1.21.4
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
protobuf==3.19.3
pycodestyle==2.8.0
pyparsing==3.0.6
python-dateutil==2.8.2
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-probability==0.15.0
termcolor==1.1.0
threadpoolctl==3.0.0
typing_extensions==4.0.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
